Create python
In the above example #1, first, we import the cv2 library as reference variable c; you can set the reference variable on its own, just easy to use to reduce the library word. The first thing needs to install the python-OpenCV library through the command line or with IDE. Copy the file location path and past it in quotes. On the right side, use a character r which means to read. c.imread(File_path) needs two parameters: file path and flag integer value. when you are using IDE such as pycharm or Spyder, use a method waitKey with parameter 0, show image in the screen.