Create python_1
In the above example #2, firstly, we import the cv2 library as reference variable c; rest of the code you can use c instead of cv2. You can set the reference variable on its own, just easy to use to reduce the library word. Copy the file location path and past it in quotes. On the right side, use a character r which means to read.c.imread(File_path) needs two parameters: file path and a flag integer value. As differentiate from example #1, you set the flag value to 0; the method read the one channel of the file (gray image).